move sidekick_plugin_installer/test/util to sidekick_test
