+++ sidekick_plugin_installer/lib/sidekick_plugin_installer.dart
@@ -1,7 +1,7 @@
 library sidekick_plugin_installer;
 
-export 'package:sidekick_plugin_installer/src/add_import.dart';
 export 'package:sidekick_plugin_installer/src/add_dependency.dart';
+export 'package:sidekick_plugin_installer/src/add_import.dart';
 export 'package:sidekick_plugin_installer/src/plugin_context.dart';
 export 'package:sidekick_plugin_installer/src/pub_add_dependency.dart';
 export 'package:sidekick_plugin_installer/src/pub_add_local_dependency.dart';
