+++ sidekick_plugin_installer/test/add_dependency_test.dart
@@ -2,9 +2,8 @@ import 'package:pubspec/pubspec.dart'
     show PathReference, HostedReference, GitReference;
 import 'package:sidekick_core/sidekick_core.dart';
 import 'package:sidekick_plugin_installer/sidekick_plugin_installer.dart';
+import 'package:sidekick_test/sidekick_test.dart';
 import 'package:test/test.dart';
-
-import 'util/local_testing.dart';
 
 void main() {
   late final Directory packageDir;
